Fix empresa, user, Appconfig, etc. (#41)
* Fix: Se corrigen e implementan las capas de entidad, DTOs, mappers, repositorios, servicios y controladores.

* Fix: se corrigen errores cruzados

* Fix: se corrigen el uso de hateoas

+++ src/main/java/com/coagronet/criterioEvaluacion/dtos/CriterioEvaluacionDTO.java
@@ -8,9 +8,9 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class CriterioEvaluacionDTO {
-    private Integer id;
-    private Integer tipoEvaluacion;
+    private Long id;
+    private Long tipoEvaluacion;
     private String nombre;
     private String descripcion;
-    private Integer estado;
+    private Long estado;
 }
